+++ javascript/modules/client/test/integration/client.test.ts
@@ -250,11 +250,33 @@ describe("Client", () => {
       const client = new Client(context);
 
       const depositParams: IDepositParams = {
-        daoAddress: daoAddress,
+        daoAddressOrEns: daoAddress,
         amount: BigInt(1234),
       };
 
       const gasFeesEstimation = await client.estimation.deposit(
+        depositParams,
+      );
+
+      expect(typeof gasFeesEstimation).toEqual("object");
+      expect(typeof gasFeesEstimation.average).toEqual("bigint");
+      expect(typeof gasFeesEstimation.max).toEqual("bigint");
+      expect(gasFeesEstimation.max).toBeGreaterThan(BigInt(0));
+      expect(gasFeesEstimation.max).toBeGreaterThan(gasFeesEstimation.average);
+    });
+    it("Should estimate gas fees updating allowance", async () => {
+      const context = new Context(contextParamsLocalChain);
+      const client = new Client(context);
+
+      const tokenContract = await deployErc20(client);
+
+      const depositParams: IDepositParams = {
+        daoAddressOrEns: daoAddress,
+        amount: BigInt(1234),
+        tokenAddress: tokenContract.address,
+      };
+
+      const gasFeesEstimation = await client.estimation.updateAllowance(
         depositParams,
       );
 
@@ -269,7 +291,7 @@ describe("Client", () => {
       const client = new Client(context);
 
       const depositParams: IDepositParams = {
-        daoAddress: daoAddress,
+        daoAddressOrEns: daoAddress,
         amount: BigInt(1234),
       };
 
@@ -277,7 +299,7 @@ describe("Client", () => {
       if (!provider) throw new Error("No provider");
 
       expect(
-        (await provider.getBalance(depositParams.daoAddress))
+        (await provider.getBalance(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("0");
 
@@ -299,7 +321,7 @@ describe("Client", () => {
       }
 
       expect(
-        (await provider.getBalance(depositParams.daoAddress))
+        (await provider.getBalance(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("1234");
     });
@@ -311,14 +333,14 @@ describe("Client", () => {
       const tokenContract = await deployErc20(client);
 
       const depositParams: IDepositParams = {
-        daoAddress: daoAddress,
+        daoAddressOrEns: daoAddress,
         amount: BigInt(5),
         tokenAddress: tokenContract.address,
         reference: "My reference",
       };
 
       expect(
-        (await tokenContract.functions.balanceOf(depositParams.daoAddress))
+        (await tokenContract.functions.balanceOf(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("0");
 
@@ -352,7 +374,7 @@ describe("Client", () => {
       }
 
       expect(
-        (await tokenContract.functions.balanceOf(depositParams.daoAddress))
+        (await tokenContract.functions.balanceOf(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("5");
     });
@@ -364,14 +386,14 @@ describe("Client", () => {
       const tokenContract = await deployErc20(client);
 
       const depositParams: IDepositParams = {
-        daoAddress: daoAddress,
+        daoAddressOrEns: daoAddress,
         amount: BigInt(7),
         tokenAddress: tokenContract.address,
         reference: "My reference",
       };
 
       expect(
-        (await tokenContract.functions.balanceOf(depositParams.daoAddress))
+        (await tokenContract.functions.balanceOf(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("0");
 
@@ -379,17 +401,17 @@ describe("Client", () => {
       expect(
         (await tokenContract.functions.allowance(
           await client.web3.getSigner()?.getAddress(),
-          depositParams.daoAddress,
+          depositParams.daoAddressOrEns,
         )).toString(),
       ).toBe("0");
 
-      await tokenContract.functions.approve(depositParams.daoAddress, 10)
+      await tokenContract.functions.approve(depositParams.daoAddressOrEns, 10)
         .then((tx) => tx.wait());
 
       expect(
         (await tokenContract.functions.allowance(
           await client.web3.getSigner()?.getAddress(),
-          depositParams.daoAddress,
+          depositParams.daoAddressOrEns,
         )).toString(),
       ).toBe("10");
 
@@ -416,7 +438,7 @@ describe("Client", () => {
       }
 
       expect(
-        (await tokenContract.functions.balanceOf(depositParams.daoAddress))
+        (await tokenContract.functions.balanceOf(depositParams.daoAddressOrEns))
           .toString(),
       ).toBe("7");
     });