+++ javascript/modules/client/CHANGELOG.md
@@ -20,6 +20,8 @@ TEMPLATE:
 ### Changed
 - Examples folder and files
 - Automatic generated README from examples
+### Added
+- Update Allowance estimation
 ## 0.11.0-alpha
 On 2022-10-17 09:26:34
 ### Fixed